pyproject.toml
[tool.poetry]
name = "tree_ml"
version = "0.2.3"
description = "Tree ML Platform for Detection and Risk Assessment"
authors = ["Tree ML Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
numpy = ">=1.23.0,<=2.1.1"
scipy = "^1.15.2"
requests = "^2.32.3"
tqdm = "^4.67.1"
pyyaml = "^6.0.2" 
pillow = "^11.1.0"
pandas = "^2.2.3"
matplotlib = "^3.10.1"
scikit-learn = "^1.6.1"
torch = "^2.6.0"
torchvision = "^0.21.0"
ultralytics = "8.3.116"
segment-anything = "^1.0"
onnx = "^1.17.0"
onnxruntime = "^1.21.0"
deepforest = "^1.5.2"
opencv-python = "^4.11.0.86"
fastapi = "^0.95.0"
flask = {extras = ["async"], version = "^3.1.0"}
flask-cors = "^4.0.0"
aiohttp = "^3.11.12"
gunicorn = "^21.2.0"
zarr = "^2.16.1"
h5py = "^3.13.0"
redis = "^5.2.1"
s2sphere = "^0.2.5"
pyproj = "^3.7.0"
rasterio = "^1.4.3"
geopandas = "^1.0.1"
laspy = {extras = ["laszip", "lazrs"], version = "^2.5.4"}
psutil = "^7.0.0"
gputil = "^1.4.0"
pynvml = "^12.0.0"
asyncpg = "^0.30.0"
google-cloud-bigquery = "^3.31.0"
aiofiles = "^24.1.0"
python-dotenv = "^1.1.0"
transformers = "^4.52.3"
uvicorn = "^0.34.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^23.12.1"
mypy = "^1.15.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
httpx = "0.24.1"
python-multipart = "^0.0.20"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88